Move classmethod_ to init
